+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Added
   - Listing webhooks
   - Viewing details of a webhook
+  - Creating webhooks
 
 ## [1.1.0] - 2019-10-07